requirements.txt
Flask==2.3.2
gspread==5.7.2
google-auth==2.17.3
gunicorn==20.1.0